@@ -1,56 +1,55 @@
 {
-	"name": "nvm-framework",
-	"version": "0.0.0",
-	"scripts": {
-		"ng": "ng",
-		"start": "ng serve",
-		"build": "ng build",
-		"test": "ng test",
-		"lint": "ng lint",
-		"e2e": "ng e2e"
-	},
-	"private": true,
-	"dependencies": {
-		"@angular/animations": "~8.2.14",
-		"@angular/common": "~8.2.14",
-		"@angular/compiler": "~8.2.14",
-		"@angular/core": "~8.2.14",
-		"@angular/forms": "~8.2.14",
-		"@angular/platform-browser": "~8.2.14",
-		"@angular/platform-browser-dynamic": "~8.2.14",
-		"@angular/router": "~8.2.14",
-		"chart.js": "^2.9.3",
-		"lodash": "^4.17.15",
-		"quagga": "^0.12.1",
-		"rxjs": "~6.4.0",
-		"rxjs-tslint": "^0.1.7",
-		"tslib": "^1.10.0",
-		"zone.js": "~0.9.1"
-	},
-	"devDependencies": {
-		"@angular-devkit/build-angular": "^0.803.21",
-		"@angular-devkit/build-ng-packagr": "^0.803.21",
-		"@angular/cli": "~8.3.21",
-		"@angular/compiler-cli": "~8.2.14",
-		"@angular/language-service": "~8.2.14",
-		"@types/node": "~8.9.4",
-		"@types/jasmine": "~3.3.8",
-		"@types/jasminewd2": "~2.0.3",
-		"@types/chart.js": "^2.9.11",
-		"@types/lodash": "^4.14.149",
-		"codelyzer": "^5.0.0",
-		"jasmine-core": "~3.4.0",
-		"jasmine-spec-reporter": "~4.2.1",
-		"karma": "~4.1.0",
-		"karma-chrome-launcher": "~2.2.0",
-		"karma-coverage-istanbul-reporter": "~2.0.1",
-		"karma-jasmine": "~2.0.1",
-		"karma-jasmine-html-reporter": "^1.4.0",
-		"protractor": "~5.4.0",
-		"ts-node": "~7.0.0",
-		"tslint": "~5.15.0",
-		"typescript": "~3.5.3",
-		"ng-packagr": "^5.7.1",
-		"tsickle": "^0.34.1"
-	}
+  "name": "nvm-framework",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~9.1.0",
+    "@angular/common": "~9.1.0",
+    "@angular/compiler": "~9.1.0",
+    "@angular/core": "~9.1.0",
+    "@angular/forms": "~9.1.0",
+    "@angular/platform-browser": "~9.1.0",
+    "@angular/platform-browser-dynamic": "~9.1.0",
+    "@angular/router": "~9.1.0",
+    "chart.js": "^2.9.3",
+    "lodash": "^4.17.15",
+    "quagga": "^0.12.1",
+    "rxjs": "~6.5.4",
+    "rxjs-tslint": "^0.1.7",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.901.0",
+    "@angular-devkit/build-ng-packagr": "~0.901.0",
+    "@angular/cli": "~9.1.0",
+    "@angular/compiler-cli": "~9.1.0",
+    "@angular/language-service": "~9.1.0",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/chart.js": "^2.9.11",
+    "@types/lodash": "^4.14.149",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.8.3",
+    "ng-packagr": "^9.0.0"
+  }
 }
