@@ -6,8 +6,12 @@ export * from './lib/nvm-autocomplete.module';
 export * from './lib/nvm-autocomplete/nvm-autocomplete.component';
 export * from './lib/nvm-autocomplete/nvm-chips/nvm-chips.component';
 export * from './lib/nvm-autocomplete/models/nvm-autocomplete-item';
+export * from './lib/nvm-autocomplete/nvm-suggestions/nvm-suggestions.component';
 export * from './lib/nvm-autocomplete/directives/nvm-autocomplete-element.directive';
 export * from './lib/nvm-autocomplete/directives/nvm-autocomplete-item-outlet.directive';
-export * from './lib/nvm-autocomplete/nvm-suggestions/nvm-suggestions.component';
 export * from './lib/nvm-autocomplete/directives/nvm-chip-content.directive';
 export * from './lib/nvm-autocomplete/directives/nvm-suggestion-content.directive';
+export * from './lib/nvm-autocomplete/directives/nvm-chip-delete.directive';
+export * from './lib/nvm-autocomplete/directives/nvm-dd-button-outlet.directive';
+export * from './lib/nvm-autocomplete/directives/nvm-dd-button.directive';
+export * from './lib/nvm-autocomplete/directives/nvm-delete-chip-outlet.directive';
